--- analysis/client/src/BarChart.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import { Bar } from 'react-chartjs-2';
-
-const BarChart = () => {
-  // Data for the chart
-  const data = {
-    labels: ['January', 'February', 'March', 'April', 'May', 'June'],
-    datasets: [
-      {
-        label: 'Sales',
-        data: [12, 19, 3, 5, 2, 3], // Example data for the bars
-        backgroundColor: 'rgba(54, 162, 235, 0.5)', // Bar color
-      },
-    ],
-  };
-
-  // Chart options
-  const options = {
-    scales: {
-      y: {
-        beginAtZero: true,
-      },
-    },
-  };
-
-  return (
-    <div>
-      <h2>Bar Chart Example</h2>
-      <Bar data={data} options={options} />
-    </div>
-  );
-};
-
-export default BarChart;
